+++ versiondb/tsrocksdb/iterator.go
@@ -21,6 +21,10 @@ type rocksDBIterator struct {
 	// rocksdb_readoptions_t struct (timestamp_ub_). Destroying ReadOptions while
 	// the iterator is alive causes a dangling pointer and use-after-free.
 	readOpts *grocksdb.ReadOptions
+
+	// err is captured from source.Err() on Close so Error() remains usable after
+	// source is nilled out.
+	err error
 }
 
 var _ versiondb.Iterator = (*rocksDBIterator)(nil)
@@ -147,19 +151,28 @@ func (itr *rocksDBIterator) timestamp() uint64 {
 func (itr *rocksDBIterator) trySkipZeroVersion() {
 	if itr.skipVersionZero {
 		for itr.Valid() && itr.timestamp() == 0 {
-			itr.Next()
+			if itr.isReverse {
+				itr.source.Prev()
+			} else {
+				itr.source.Next()
+			}
 		}
 	}
 }
 
 // Error implements Iterator.
 func (itr *rocksDBIterator) Error() error {
+	if itr.source == nil {
+		return itr.err
+	}
 	return itr.source.Err()
 }
 
 // Close implements Iterator.
 func (itr *rocksDBIterator) Close() error {
+	itr.isInvalid = true
 	if itr.source != nil {
+		itr.err = itr.source.Err()
 		itr.source.Close()
 		itr.source = nil
 	}
@@ -167,7 +180,7 @@ func (itr *rocksDBIterator) Close() error {
 		itr.readOpts.Destroy()
 		itr.readOpts = nil
 	}
-	return nil
+	return itr.err
 }
 
 func (itr *rocksDBIterator) assertIsValid() {